requirements.txt
asgiref==3.7.1
Django==3.2.19
django-tenants==3.5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4